Fix pull to refresh
