+++ ts/features/messages/components/Home/__tests__/MessageList.test.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { RefreshControlProps } from "react-native";
 import { createStore } from "redux";
 import { fireEvent } from "@testing-library/react-native";
 import { MessageList } from "../MessageList";
@@ -9,9 +10,11 @@ import { preferencesDesignSystemSetEnabled } from "../../../../../store/actions/
 import { renderScreenWithNavigationStoreContext } from "../../../../../utils/testWrapper";
 import { MESSAGES_ROUTES } from "../../../navigation/routes";
 import * as homeUtils from "../homeUtils";
-import { loadNextPageMessages } from "../../../store/actions";
+import {
+  loadNextPageMessages,
+  reloadAllMessages
+} from "../../../store/actions";
 import { pageSize } from "../../../../../config";
-import { RefreshControlProps } from "../CustomRefreshControl";
 
 const mockDispatch = jest.fn();
 jest.mock("react-redux", () => ({
@@ -67,7 +70,87 @@ describe("MessageList", () => {
 
     expect(mockDispatch.mock.calls.length).toBe(0);
   });
-  it("should have the custom refresh control", () => {
+  it("should dispatch 'reloadAllMessages.request' when output from 'getReloadAllMessagesActionForRefreshIfAllowed' is not undefined, INBOX", () => {
+    const expectedCategory: MessageListCategory = "INBOX";
+    const expectedAction = reloadAllMessages.request({
+      pageSize,
+      filter: { getArchived: false }
+    });
+    jest
+      .spyOn(homeUtils, "getReloadAllMessagesActionForRefreshIfAllowed")
+      .mockImplementation((_state, category) =>
+        category === expectedCategory ? expectedAction : undefined
+      );
+
+    const component = renderComponent(expectedCategory);
+    const messageList = component.getByTestId("message_list_inbox");
+    expect(messageList).toBeTruthy();
+
+    const { refreshControl } = messageList.props;
+    expect(refreshControl).toBeTruthy();
+
+    const { onRefresh } = refreshControl.props as RefreshControlProps;
+    expect(onRefresh).toBeTruthy();
+
+    onRefresh!();
+
+    expect(mockDispatch.mock.calls.length).toBe(1);
+    expect(mockDispatch.mock.calls[0][0]).toStrictEqual(expectedAction);
+  });
+  it("should dispatch 'reloadAllMessages.request' when output from 'getReloadAllMessagesActionForRefreshIfAllowed' is not undefined, ARCHIVE", () => {
+    const expectedCategory: MessageListCategory = "ARCHIVE";
+    const expectedAction = reloadAllMessages.request({
+      pageSize,
+      filter: { getArchived: true }
+    });
+    jest
+      .spyOn(homeUtils, "getReloadAllMessagesActionForRefreshIfAllowed")
+      .mockImplementation((_state, category) =>
+        category === expectedCategory ? expectedAction : undefined
+      );
+
+    const component = renderComponent(expectedCategory);
+    const messageList = component.getByTestId("message_list_archive");
+    expect(messageList).toBeTruthy();
+
+    const { refreshControl } = messageList.props;
+    expect(refreshControl).toBeTruthy();
+
+    const { onRefresh } = refreshControl.props as RefreshControlProps;
+    expect(onRefresh).toBeTruthy();
+
+    onRefresh!();
+
+    expect(mockDispatch.mock.calls.length).toBe(1);
+    expect(mockDispatch.mock.calls[0][0]).toStrictEqual(expectedAction);
+  });
+  it("should not dispatch 'reloadAllMessages.request' when output from 'getReloadAllMessagesActionForRefreshIfAllowed' is undefined", () => {
+    const expectedCategory: MessageListCategory = "INBOX";
+    const expectedAction = reloadAllMessages.request({
+      pageSize,
+      filter: { getArchived: false }
+    });
+    jest
+      .spyOn(homeUtils, "getReloadAllMessagesActionForRefreshIfAllowed")
+      .mockImplementation((_state, category) =>
+        category === expectedCategory ? undefined : expectedAction
+      );
+
+    const component = renderComponent(expectedCategory);
+    const messageList = component.getByTestId("message_list_inbox");
+    expect(messageList).toBeTruthy();
+
+    const { refreshControl } = messageList.props;
+    expect(refreshControl).toBeTruthy();
+
+    const { onRefresh } = refreshControl.props as RefreshControlProps;
+    expect(onRefresh).toBeTruthy();
+
+    onRefresh!();
+
+    expect(mockDispatch.mock.calls.length).toBe(0);
+  });
+  it("should have the refresh control set", () => {
     const expectedCategory: MessageListCategory = "INBOX";
     const component = renderComponent(expectedCategory);
     const messageList = component.getByTestId("message_list_inbox");
@@ -76,9 +159,9 @@ describe("MessageList", () => {
     const { refreshControl } = messageList.props;
     expect(refreshControl).toBeTruthy();
 
-    expect((refreshControl.props as RefreshControlProps).category).toBe(
-      expectedCategory
-    );
+    expect(
+      (refreshControl.props as RefreshControlProps).onRefresh
+    ).toBeTruthy();
   });
 });
